requirements.txt
fastapi
uvicorn
phidata
phidata[aws]
openai
duckduckgo-search
pymongo
pgvector
sqlalchemy
pgvector
psycopg